fake-backends/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
python-multipart==0.0.12
jinja2==3.1.4
requests==2.32.0
pyzeebe==3.0.3
mcp==1.1.2
httpx==0.27.0
sse-starlette==2.1.3
